+++ jdaviz/configs/default/tests/test_data_menu.py
@@ -24,7 +24,7 @@ def test_data_menu_toggles(specviz_helper, spectrum1d):
     assert len(dm._obj.visible_layers) == 1
 
     # add a subset and make sure it appears for the first data entry but not the second
-    specviz_helper.plugins['Subset Tools']._obj.import_region(
+    specviz_helper.plugins['Subsets'].import_region(
         SpectralRegion(6000 * spectrum1d.spectral_axis.unit, 6500 * spectrum1d.spectral_axis.unit))
 
     assert len(dm._obj.layer_items) == 3
@@ -119,14 +119,14 @@ def test_data_menu_remove_subset(specviz_helper, spectrum1d):
     specviz_helper.load_data(new_spec, data_label="test2")
 
     dm = specviz_helper.viewers['spectrum-viewer']._obj.data_menu
-    sp = specviz_helper.plugins['Subset Tools']
+    sp = specviz_helper.plugins['Subsets']
 
-    sp._obj.import_region(SpectralRegion(6000 * spectrum1d.spectral_axis.unit,
-                                         6100 * spectrum1d.spectral_axis.unit),
-                          combination_mode='new')
-    sp._obj.import_region(SpectralRegion(6000 * spectrum1d.spectral_axis.unit,
-                                         6100 * spectrum1d.spectral_axis.unit),
-                          combination_mode='new')
+    sp.import_region(SpectralRegion(6000 * spectrum1d.spectral_axis.unit,
+                                    6100 * spectrum1d.spectral_axis.unit),
+                     combination_mode='new')
+    sp.import_region(SpectralRegion(6000 * spectrum1d.spectral_axis.unit,
+                                    6100 * spectrum1d.spectral_axis.unit),
+                     combination_mode='new')
 
     assert dm.layer.choices == ['test', 'test2', 'Subset 1', 'Subset 2']
     dm.layer.selected = ['Subset 1']
@@ -153,10 +153,10 @@ def test_data_menu_subset_appearance(specviz_helper, spectrum1d):
     specviz_helper.load_data(spectrum1d, data_label="test")
 
     dm = specviz_helper.viewers['spectrum-viewer']._obj.data_menu
-    sp = specviz_helper.plugins['Subset Tools']
+    sp = specviz_helper.plugins['Subsets']
 
-    sp._obj.import_region(SpectralRegion(6000 * spectrum1d.spectral_axis.unit,
-                                         6100 * spectrum1d.spectral_axis.unit))
+    sp.import_region(SpectralRegion(6000 * spectrum1d.spectral_axis.unit,
+                                    6100 * spectrum1d.spectral_axis.unit))
 
     assert dm.layer.choices == ['test', 'Subset 1']
 
@@ -169,14 +169,14 @@ def test_data_menu_view_info(specviz_helper, spectrum1d):
 
     dm = specviz_helper.viewers['spectrum-viewer']._obj.data_menu
     mp = specviz_helper.plugins['Metadata']
-    sp = specviz_helper.plugins['Subset Tools']
+    sp = specviz_helper.plugins['Subsets']
 
-    sp._obj.import_region(SpectralRegion(6000 * spectrum1d.spectral_axis.unit,
-                                         6100 * spectrum1d.spectral_axis.unit),
-                          combination_mode='new')
-    sp._obj.import_region(SpectralRegion(6200 * spectrum1d.spectral_axis.unit,
-                                         6300 * spectrum1d.spectral_axis.unit),
-                          combination_mode='new')
+    sp.import_region(SpectralRegion(6000 * spectrum1d.spectral_axis.unit,
+                                    6100 * spectrum1d.spectral_axis.unit),
+                     combination_mode='new')
+    sp.import_region(SpectralRegion(6200 * spectrum1d.spectral_axis.unit,
+                                    6300 * spectrum1d.spectral_axis.unit),
+                     combination_mode='new')
 
     assert dm.layer.choices == ['test', 'test2', 'Subset 1', 'Subset 2']
 
@@ -186,7 +186,7 @@ def test_data_menu_view_info(specviz_helper, spectrum1d):
 
     dm.layer.selected = ["Subset 2"]
     dm.view_info()
-    assert sp._obj.subset_select.selected == "Subset 2"
+    assert sp.subset.selected == "Subset 2"
 
     dm.layer.selected = ["test", "test2"]
     with pytest.raises(ValueError, match="Only one layer can be selected to view info"):