+++ jdaviz/configs/imviz/plugins/viewers.py
@@ -81,7 +81,7 @@ class ImvizImageView(JdavizViewerMixin, BqplotImageView, AstrowidgetsImageViewer
             key_pressed = data['key']
 
             if key_pressed in ('b', 'B'):
-                self.blink_once(reversed=key_pressed=='B')  # noqa: E225
+                self.blink_once(reversed=key_pressed == 'B')
 
     def blink_once(self, reversed=False):
         # Simple blinking of images - this will make it so that only one
@@ -190,7 +190,7 @@ class ImvizImageView(JdavizViewerMixin, BqplotImageView, AstrowidgetsImageViewer
         When `True`, it sets the coords, otherwise it resets.
 
         ``reverse=True`` is only for internal roundtripping (e.g., centroiding
-        in Subset Tools plugin). Never use this for coordinates display panel.
+        in Subsets plugin). Never use this for coordinates display panel.
 
         """
         # By default we'll assume the coordinates are valid and within any applicable bounding box.