+++ jdaviz/configs/imviz/tests/test_delete_data.py
@@ -29,11 +29,11 @@ class TestDeleteData(BaseImviz_WCS_WCS):
 
         # Add a subset
         reg = CirclePixelRegion(PixCoord(2, 2), 3).to_sky(self.wcs_1)
-        self.imviz.plugins['Subset Tools']._obj.import_region(reg)
+        self.imviz.plugins['Subsets'].import_region(reg)
 
-        self.imviz.plugins['Subset Tools']._obj.combination_mode.selected = 'new'
+        self.imviz.plugins['Subsets'].combination_mode = 'new'
         reg = RectanglePixelRegion(PixCoord(1, 1), 2, 2).to_sky(self.wcs_1)
-        self.imviz.plugins['Subset Tools']._obj.import_region(reg)
+        self.imviz.plugins['Subsets'].import_region(reg)
 
         assert len(self.imviz.app.data_collection.subset_groups) == 2
 
@@ -81,6 +81,7 @@ class TestDeleteData(BaseImviz_WCS_WCS):
 
 class TestDeleteWCSLayerWithSubset(BaseImviz_WCS_GWCS):
     """Regression test for https://jira.stsci.edu/browse/JDAT-3958"""
+
     def test_delete_wcs_layer_with_subset(self):
         lc_plugin = self.imviz.plugins['Orientation']
         lc_plugin.align_by = 'WCS'
@@ -91,7 +92,7 @@ class TestDeleteWCSLayerWithSubset(BaseImviz_WCS_GWCS):
         # Create a rotated ellipse.
         reg = EllipsePixelRegion(
             PixCoord(3.5, 4.5), width=2, height=5, angle=Angle(30, 'deg')).to_sky(self.wcs_1)
-        self.imviz.plugins['Subset Tools']._obj.import_region(reg)
+        self.imviz.plugins['Subsets'].import_region(reg)
 
         # Switch back to Default Orientation.
         self.imviz.app._change_reference_data("Default orientation")