+++ CHANGES.rst
@@ -26,6 +26,8 @@ Specviz2d
 API Changes
 -----------
 - Removed API access to plugins that have passed the deprecation period: Links Control, Canvas Rotation, Export Plot. [#3270]
+  
+- Renamed the ``Subset Tools`` plugin to ``Subsets`` which now exposes the ``subset``, ``combination_mode``, ``get_center``, and ``set_center`` in the user API. [#3293]
 
 Cubeviz
 ^^^^^^^